+++ tests/tde/theme/icons.lua
@@ -278,9 +278,13 @@ function Test_theme_icon_dir()
     assert(file_exists(icons.dir), "Check that the dir icon exists")
 end
 
+function Test_theme_icon_developer()
+    assert(file_exists(icons.developer), "Check that the dir icon exists")
+end
+
 
 function Test_that_all_icons_are_tested()
-    local amount = 63
+    local amount = 64
     assert(
         tablelength(icons) == amount,
         "It looks like you forgot to add some icons to the test suite, add them and then update this test (you forgot: " ..